创建数据库

+++ xiucailv/src/main/java/com/sunveee/xiucailv/dao/UserMapper.java
@@ -6,13 +6,13 @@ import com.sunveee.xiucailv.domain.User;
 import com.sunveee.xiucailv.util.PageEntity;
 
 public interface UserMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(User record);
 
     int insertSelective(User record);
 
-    User selectByPrimaryKey(Integer id);
+    User selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(User record);
