+++ xiucailv/src/main/java/com/sunveee/xiucailv/service/UserService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.sunveee.xiucailv.domain.User;
 
 public interface UserService {
-    public User getUserById(int userId);
+    public User getUserById(Long userId);
 
     public List<User> getUserPage(int pageNo, int pageSize);
